+++ packages/react-server/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+
+- Move default options from webpack plugin into react-server. ([#1514](https://github.com/Shopify/quilt/pull/1514))
 
 ## [0.15.0] - 2020-06-06